# src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Galactic War API for FAForever"
  version: "1.0.0"
  title: "Galactic War"
  contact: {}
  license:
    name: "Apache License 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
host: "gw.faforever.com"
basePath: "/1.0"
...
schemes:
- "http"
paths:
...
  /characters/hall_of_fame/{faction}:
    get:
      tags:
      - "Character management"
      summary: "Retrieve the hall of fame-character by faction"
      consumes: []
      produces:
      - "application/model"
      parameters:
      - name: "faction"
        in: "path"
        description: "Can be 'aeon', 'cybran', 'seraphim', 'uef' or 'all'"
        required: true
        type: "string"
      responses:
        200:
          description: "Succsess"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
...
  /characters/{id}/credit_journal:
    get:
      tags:
      - "Character management"
      summary: "Retrieve credit journal"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        description: "id of the character"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Success"
        404:
          description: "Status 404 - Player not found"
...
definitions:
...
  Character:
    type: "object"
    required:
    - "faction"
    - "id"
    - "killedAt"
    - "killedBy"
    - "name"
    - "rankId"
    - "rankTitle"
    properties:
      currentCredits:
        type: "integer"
        format: "int64"
      currentXp:
        type: "integer"
        format: "int64"
      faction:
        type: "string"
      id:
        type: "integer"
        format: "int32"
      killedAt:
        type: "string"
        format: "date"
        description: "can be NULL"
      killedBy:
        description: "can be NULL"
        $ref: "#/definitions/Character"
      name:
        type: "string"
      rankId:
        type: "integer"
        format: "int32"
        description: "Internal ID of the character's rank"
      rankTitle:
        type: "string"
        description: "Textual representation of the character's rank"
...